+++ src/GRA.Controllers/ViewModel/Profile/HouseholdAddViewModel.cs
@@ -21,5 +21,8 @@ namespace GRA.Controllers.ViewModel.Profile
         [DisplayName("Is this your first time participating?")]
         [Required(ErrorMessage = "Please let us know if this is your first time participating in the program")]
         public string IsFirstTime { get; set; }
+
+        public string TranslationDescriptionPastTense { get; set; }
+        public string ActivityDescriptionPlural { get; set; }
     }
 }