+++ src/GRA/Defaults.cs
@@ -2,6 +2,7 @@
 {
     public static class Defaults
     {
+        public const int MaxDailyActivityGoal = 100000;
         public const string SiteLogoPath = "~/images/great-reading-adventure-logo@1x.png";
     }
 }