+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Setting for maximum allowed points earned total
 - Private and chater school district types
 - Homeschool and school not listed options
+- Setting for asking participants daily personal goals
 
 ### Changed
 - Participant mail subject to be clickable